refactor(content-engine): split loader.rs by row family (#244)
loader.rs was 1094 lines (over the 700-line cap). The file was three
big match-arm converters (trigger, condition, action) plus orchestration;
each converter grows independently as new variants land. Split into a
directory with one converter per file:

  loader/mod.rs      — load_chains_from_json, build_chains_from_rows,
                       public DbXRow structs (orchestration + public API)
  loader/trigger.rs  — convert_trigger
  loader/condition.rs — convert_condition + parse_comparison_op,
                        parse_mission_status, parse_step_status
  loader/action.rs   — convert_action + parse_destination
  loader/tests.rs    — existing test suite, unchanged behavior

Public surface preserved: external callers
(`services::cell::content::engine_loader`) still import via
`cimmeria_content_engine::loader::{build_chains_from_rows, DbXRow}`.

Sub-module helpers are `pub(super)` so the sibling tests module can
exercise them through `super::trigger::convert_trigger` etc.

All 17 loader tests pass; clippy clean (with the doc-list indentation
fix); cargo fmt applied.
